requirements.txt
streamlit
pandas
seaborn
yfinance
matplotlib
japanize_matplotlib
jupyterlab
pyarrow
google-cloud-bigquery
google-cloud-storage
db-dtypes
tqdm
beautifulsoup4
xlrd
